requirements.txt
azure-storage-blob==12.5.0
azure-identity==1.4.1
azure-storage-file-datalake==12.1.2
click==7.1.2
azure-appconfiguration==1.1.1
